check if file exists before delete

+++ sources/base.sh
@@ -24,11 +24,13 @@ function safe-rm()
 {
     local file="$1"
 
-    which shred > /dev/null
-    if [ $? -eq 0 ]; then
-        # if available, use shred to safelly delete the key file
-        shred -fu "$file"
-    else
-        rm -f "$file"
+    if [ -f $file ]; then
+        which shred > /dev/null
+        if [ $? -eq 0 ]; then
+            # if available, use shred to safelly delete the key file
+            shred -fu "$file"
+        else
+            rm -f "$file"
+        fi
     fi
 }
